Android support and minor fix
- Unity version updated to 2017.2.0f3
- Android Studio Project (.so files to be were currently set at the
Unity project are built to
AndroidProj\unityforcpp\.externalNativeBuild\cmake\debug\obj)
- minor changes on cs and cpp sources for Android compilation
- minor bug fix at UnityArray
- .dll plugin binary compiled for x64

+++ UnityForCppPlugin/Source/Shared.h
@@ -26,8 +26,8 @@ typedef int16_t int16; //short
 typedef uint16_t uint16; //unsigned short
 typedef int32_t int32; //int
 typedef uint32_t uint32; //unsigned int
-typedef int64_t int64; //long long
-typedef uint64_t uint64; //unsigned long long
+typedef long long int64; //long long
+typedef unsigned long long uint64; //unsigned long long
 
 
 #include "stdio.h"
